Support byte_container_util on non-x86 platforms
The implementation of SerializeByteContainers relied on the fact that
the destination platform was little-endian. The implementation was not
generalized to work on non-x86 platforms as Asylo's POSIX layer did not
yet support endian.h. Now that Asylo supports endianness conversions,
the implementation can be updated to work on platforms that are not
little-endian.

PiperOrigin-RevId: 285255830
Change-Id: I05fd2902a6f5af278c49a75c69f674f3741422cd