Remove AttestationDomainClient and usages
This change removes the AttestationDomain client, which is now obsolete
because the AttestationDomainService has been removed.

The related code in EnclaveManager has been removed. It is no longer a
requirement to configure the EnclaveManager with
EnclaveManager::Configure() before obtaining the static instance. The
Configure() method is now deprecated and will be removed in a future
release.

The HostConfig proto field in EnclaveConfig is now also deprecated.
Related methods for creating configs in enclave_config_util.h have been
updated so that they do not require a HostConfig parameter.

PiperOrigin-RevId: 285320320
Change-Id: I63f1a5377439cab0b29024d893253a87b1dcaff6