Add multithread fork test
The left-over issue of fork being not thread-safe is fixed by recent
changes. Now fork() waits to confirm all other threads are waiting in a
controllable state without modifying memory/holding locks both inside
and outside the enclave before proceed to fork, in both sgx hardware and
simulation mode. This change adds test that fork() is invoked while
other threads are randomly allocating/freeing/exiting/reentering the
enclave, check for fork() succeeded and checks the variables allocated
on heap and stack in the parent holds the same value in the child after
fork().

PiperOrigin-RevId: 285443851
Change-Id: Ief8b6b6c60e4708e5d5d7f86189fc162878a0db4