Make sys_futex ocalls to be host calls
Converts the untrusted futex calls to host calls, so that
they have a backend agnostic implementation. The handlers for
futex calls are made to always use a local exit call provider,
both in the presence and absence of the remote backend. This
ensures that thread waiting and waking using untrusted futex
(or in higher level abstractions using untrusted futex) always
happens locally.

PiperOrigin-RevId: 287595198
Change-Id: I34ac4689a28fdd898dbbca768b8dd59b374dee9b