Add AGE configs for FLC and non-FLC environments
Access to the PROVISION_KEY is permitted on hardware that supports
Flexible Launch Control (FLC).

Real deployments of the AGE will require access to the PROVISION_KEY.
The remote_assertion_generator_enclave_config now captures this
configuration, and remote_assertion_generator_enclave.so is the version
of the AGE that uses this configuration. This AGE can be used in tests
and environments in which FLC support is required.

None of the existing AGE tests require access the AGE to access the
PROVISION_KEY so these tests use
remmote_assertion_generator_enclave_no_flc.so, which is based on
remote_assertion_generator_enclave_no_flc_config (formerly
remote_assertion_generator_enclave_config).

PiperOrigin-RevId: 288072827
Change-Id: I165074275567a1efac506857a5b5a4cf7d604609