Add setters for Attributes to attributes_util
This commit moves the TestAttribute and GetPrintableAttributeList
functions from `secs_attributes` to `attributes_util` (because they
operate on the Attributes message as opposed to SecsAttributeSet),
and also adds two new functions for modifying bits of an Attributes
proto to `attributes_util` (SetAttributeBit and ClearAttributeBit).

In addition, this commit modifies the `==` operator overload on
Attributes to explicitly compare the `flags` and `xfrm` fields of the
proto, instead of depending on protobuf::util::MessageDifferencer.

PiperOrigin-RevId: 288578123
Change-Id: I11af562e9e214ddf1644cbc1783554d7c47a5070