Make attributes_util library public
This commit migrates `attributes_util` and `attributes.proto` to
`identity/platform/sgx` and changes the former's visibility to public.
It also migrates references within Asylo to the new file locations,
and leaves a forwarding header since the proto was already public.

This change exposes the new AttributeBit getters/setters on the
Attributes message, which allows users to set custom ATTRIBUTES on
SgxIdentity messages.

PiperOrigin-RevId: 288942589
Change-Id: I7d2c18cdfe01a11bcec43b6b792af48a0868bee2