Add end-to-end test for Intel ECDSA QE assertions
This test loads up a test enclave which calls into the Intel ECDSA QE
assertion generator. The assertion is passed back up to the untrusted
test code which does some basic validation.

In the future, this test will be expanded to verify the assertions with
the yet to be authored Intel ECDSA QE assertion verifier.

PiperOrigin-RevId: 289717438
Change-Id: I98dae4ab9a3e11c8df7c5d0aff226b36795b6961