Change "transitions supported" logic
This has no semantic change.

The current way of checking for whether transitions are supported is
to check package-by-package which is supported. Instead this changes
the logic to check which users are explicitly not supported. Still in
Bazel we have issues with transitions support, so all packages are still
considered unsupported.

PiperOrigin-RevId: 289731577
Change-Id: I328731a592252dcc8a3e431a64327ddfdd4b8c5c