Import Bazel macros from rules_* repo
Bazel is moving away from having native language-specific rules.
Instead, the rules need to be imported from supporting repos.

This will be required when
incompatible_load_{cc,java,proto,python}_rules_from_bzl is flipped in
Bazel, which is currently slated for Bazel 3.0.

Not all of Asylo's dependencies are yet compatible with these Bazel
flags, so we cannot yet test with them turned on to prevent backsliding.
To prevent needing to continually rebase such a large patchset, though,
we can make the bulk of these changes now and follow up with a cleanup
for any new violations of it once our dependencies are ready.

PiperOrigin-RevId: 289961936
Change-Id: I84823728ea8c66e6e60f9724cc17887deb7735ff