Add utilities to create fake AK certificate chain
This change adds utilities for creating fake certificate chains for
attestation keys wielded by the Assertion Generator Enclave (AGE).

These certificates chains are rooted in Asylo's Fake SGX PKI and,
consequently, are only intended for testing purposes.

The intention is to use these new utilities in tests for the AGE, as
well as in test deployments of the AGE.

PiperOrigin-RevId: 290159655
Change-Id: Ic04932b546a459bf6aa1b9ac40e937222ce0de64