Change Asylo signal reset logic
Change Asylo signal manager to hold a fixed size array of ResetStatus to
record the reset status of each signal. If a handler is registered with
SA_RESETHAND, this signal will be recorded as "TO_BE_RESET", otherwise
"NO_RESET". When a signal enters the enclave, if the status is
"TO_BE_RESET", it changes it to "RESET" and invoke the handler. If the
status is "RESET", it exits immediately without invoking the handler.
The reason of the change is that previously we are removing the signal
handler from the map in the signal handler to reset, which results in
heap allocation, and could cause deadlock inside a signal handler.

PiperOrigin-RevId: 292448274
Change-Id: Icf7b82cfb8094e54cb7239eb0c61c0b4bbddca0b