Mask signals during rwlock in SGX sdk
The rwlock in ecall/ocall while entering/exiting the enclave in SGX sdk
is also acquired by Asylo signal handling. If the thread is interrupted
while it's half way acquiring/releasing the lock, it could cause
undefined behavior, including deadlock or crash. Recursive mutex or read
locks do not prevent it either. So we mask all the signals during
acquring/releasing the locks.

PiperOrigin-RevId: 293022030
Change-Id: I81a7d4703e94d0dfb2950c28ec1725eb2786fab8