Fix SgxIdentityUtilTest order-dependence
Currently, if SgxIdentityUtilTest.GetSelfSgxIdentity is not the first
test to be executed in the test suite, it will fail due to attempting
to enter the new FakeEnclave while already being inside of an enclave.
This commit fixes this order-dependence by making an ExitEnclave call
if GetCurrentEnclave reports we are currently in an enclave.

PiperOrigin-RevId: 293174266
Change-Id: I99d7943ffcfff1a129d6f068124dc4257291b8bb