Fix the flakiness in signal handling in sim mode
Currently we are using the TCS active state to determine whether enter
the enclave to handle the signal, or invoke the signal handler inside
the enclave directly in SGX simulation mode, when an enclave registered
signal arrives. However, there's a tiny window between when the state is
set and when the tls are switched, which causes flakiness in signal
handling in SGX simulation mode. This commit fixes the issue by
providing the following changes:
1. When entering the enclave, mask the signals during the whole window
   while SGX simulation mode switch tls, set state and registers.
2. Use a thread local variable address to decide whether the thread
   locals are inside the enclave or not, instead of the tcs active
   state. If thread locals are inside the enclave, invoke the handler
   directly, otherwise, enter the enclave to handle the signal.
3. When exiting the enclave, we can't mask signals as it's the trusted
   side. Move the order so that it sets the state prior to switch the
   tls. Otherwise if a signal arrives after switching the tls and before
   setting the state, we will enter the enclave to handle the signal as
   the tls is already on the untrusted side, however the state is still
   TCS active so the entry will be rejected.

PiperOrigin-RevId: 294547588
Change-Id: I947db5cedc57494114fd30a8217ca77f002f2d95