Add a cc_binary that allows launching the AGE
This change adds a cc_binary that launches the Assertion Generator
Enclave and starts its attestation server. There are two modes in which
the AGE can be started:

1. Real certification: Starts the AGE with a real PCK certificate chain
provided on the command line. In this mode, the AGE's key is certified
by a real PCE. The PCE binary is expected to be specified with an
additional flag, or otherwise must be present in the filesystem in the
default location as expected by DCAP. This mode is intended for users
with access to SGX hardware that supports Flexible Launch Control (FLC).

2. Fake certification: Starts the AGE with a fake PCK certificate chain
based on Asylo's Fake SGX PKI. In this mode, a PCE is not launched. This
mode is intended completely for test setups where do not have access to
SGX hardware and only have access to SGX simulation.

PiperOrigin-RevId: 295801944
Change-Id: I5bd8fb606a435bba137c809012efab978eec5e70