Implement xattr host calls in Asylo
These calls sets/gets/lists the extended attributes for a path. This
change implements the corresponding host calls (enc_untrusted_foo) by
delegating to the host directly.
Similar as other file related syscalls, the results from these calls are
untrusted. It's not introducing new security issues, as setting/getting
attributes in "user" namespace is on the same security level as reading
or writing from a file. In both cases the results are untrusted. For
security/system/trusted namespace, the results of the security features
on the OS are untrusted, but anything on the host is already considered
untrusted inside an enclave.

PiperOrigin-RevId: 296308533
Change-Id: I8f1094007b0efbea0ed0a16bd4b68ca8369106bd