Fix read of uninitialized bytes in a unit test
The intent of TYPED_TEST(TypedBytesTest, EqualsNegative1) is to test
that a comparison fails when two buffers have different sizes.
However, the test was previously using a fixed-size type as the
input. In addition, the fixed-size buffer was not fully initialized,
causing a read of uninitialized memory in Bytes::Equal.

PiperOrigin-RevId: 297384162
Change-Id: If34b5b7c597383c6a389532123a2aca46bbc5645