Remove setsid host call test
setsid may fail on host with EPERM if the user is already the process
group leader.
Properly testing setsid needs to deal with process group hierarchy. And
since this call is directly delegated to the host, the only thing that
needs to be tested is the parameter and return result serialization and
deserialization across the enclave boundary, which is fully tested with
all other components in this test already.

PiperOrigin-RevId: 304486667
Change-Id: Ie010eabda76c5cea7d732e7cff15aa2660fb9a3e