Update FakePce::SignReport to sign REPORT.body
This fixes a bug in FakePce::SignReport. To match the behavior of the
PCE, it should be signing the REPORT.body, not the entire REPORT.

The AttestationKeyCertificateImpl makes the same incorrect assumption.
The AGE uses AttestationKeyCertificateImpl to validate certificate
chains. Add a bool to control this behavior and set this to false by
default. The age_main_test should flip this flag once the signature
verification logic in AttestationKeyCertificateImpl is updated.

PiperOrigin-RevId: 304711370
Change-Id: Ie51c6c156c2c4338df6a104f6263d9da87f39bac