Don't access UntrustedCacheMalloc on switched heap
UntrustedCacheMalloc is maintaining a table on enclave heap to store the
buffer list. If the table is modified on switched heap, it will cause
issue later when UntrustedCacheMalloc tries to free the table entries.
Now always invoke UntrustedLocalAlloc if on switched heap.
This causes two extra enclave entry/exits for enclave exit calls on
switched heap. However, there are limited enclave exits on switched
heap, and switched heap is restricted to be used in fork(2), so it won't
cause big performance issues.

PiperOrigin-RevId: 305387443
Change-Id: I8fb5bbb2dca7716d5396bbd055e372fd86780f69