Add option to age_main to use random PPIDEK
This commit adds a `--use-random-ppidek` flag to age_main. This will
generate a random RSA decryption key, which is passed along to the
PrintSgxPlatformInfo function, allowing it to print both the encrypted
and decrypted PPIDs.

It also updates the output format of the printed platform information
to a textproto to make it easier to interface with other scripts.

PiperOrigin-RevId: 305719467
Change-Id: Iba68872510e81585a785da173fd9b3ea6eaaa798