Move check for QE identity expectation in config
This commit moves the check for the QE identity expectation in the
authority config to the config verifiers library. This is the correct
place for this check because all the validation checks take place in
this library.

It also updates the library for creating assertion authority configs to
accept a provided QE identity expectation, or use the default one
exposed in Asylo.

PiperOrigin-RevId: 306724267
Change-Id: Ic27e3857543c05cf53e3189397f053865e4dc0c2