Count enclave exits in fork block algorithms
Currently fork() blocks enclave entries, and waits till all active
entries are blocked. However, if a thread exits the enclave and stays
on the untrusted side, it will not be caught by blocked entries and
fork() will timeout. This change counts the enclave exits as well to
confirm that all threads are either blocked in the enclave or staying
on the untrusted side (and will be blocked if they try to reenter)
before fork() starts to take the snapshot.
The exit count is increased everytime an ocall exits the enclave, and
decreased when they reenter, before they are blocked. fork() blocks
entries first, then count whether the blocked entries plus the exit
count matches the entry count. Therefore it should always be safe:
1) If fork() counts when a thread is still on the untrusted side, this
   thread is considered exited the enclave and won't affect the
   snapshot. If it reenters later, it will be blocked since fork()
   blocks entries prior to counting;
2) If fork() counts after a thread enters the enclave, but prior to the
   blocker, it is considered neither exited or blocked, fork() will wait
   and count again later, until it executed to the blocker code;
3) If fork() counts after a thread enters the enclave and get blocked,
   it is considered not exited, but blocked, so it's safe to continue
   taking the snapshot.

PiperOrigin-RevId: 307123902
Change-Id: Ib81b0df372b0c362d6845e0921ed06885e7ca15f