Remove safe_integer_cast library
This library was intended to be used in ASN.1 code, but that code was
refactored to not need safe integer casting. This change deletes the
safe integer casting library to clean out dead code. If a use is found
for this library in the future, then it can be restored.

PiperOrigin-RevId: 308689603
Change-Id: Idd443b362c809cd8f02f0abfc7267565301b3ce9