Fix some race conditions with detached flag
Make the detached flag atomic.

There could be threads waiting on the detached state change. We
should therefore set the state prior to signaling the conditional
variable those other threads are waiting on. Else, there is a very
small race wherein the waiting thread is woken up but the detached
state has not yet been set.

PiperOrigin-RevId: 309820054
Change-Id: I18ae24b4f0baa47b90e1f90e4abbfcfe5eee3a1c