Add LVI mitigation patch into Asylo toolchain
This commit adds a patch step into the toolchain installation process
that integrates an upstream change to Binutils provided by Intel that
implements mitigations for LVI attacks (CVE-2020-0551) [1].

Some minor changes that were not explicitly a part of the commit were
added into the patch in order to allow the commit to apply properly to
the somewhat outdated version of Binutils that Asylo depends on.

The reason that Binutils was not simply bumped was because a)
attempting to upgrade Binutils caused a cascade of other build and
toolchain breakages, and b) this commit is currently not included in a
stable release of Binutils (and the timeline of when there will be
another release is ambiguous).

[1]: https://github.com/bminor/binutils-gdb/commit/ae531041c7c5956672

PiperOrigin-RevId: 309990438
Change-Id: Id60a0d894c178305cdecfe5bd74a9f576613bf20