Add transitive LVI mitigation features
This commit implements "transitive features" within Asylo's Bazel build
system, which allows features to be specified for a single BUILD target
and subsequently propagated to all parent dependencies.

The primary purpose for this is to provide a way for Asylo users to
specify varying levels of LVI mitigations when building enclaves.
Intel's upstream mitigations come in three flavours: "All-Loads",
"Control-Flow", and "No-Auto". By default, Asylo will build enclave
targets using the most aggressive mitigation level, defined by the
"lvi_all_loads_mitigation" feature.

To override this, either pass one of the two other mitigation features
("lvi_control_flow_mitigation" or "lvi_no_auto_mitigation") to the list
of `transitive_features` for the unsigned enclave of a BUILD target, or
pass `--features="lvi_*_mitigation"` as a flag to the Bazel command.

PiperOrigin-RevId: 310599602
Change-Id: I71440ccf62d04de5538fc4656c42e268e7661246