Don't pass transitive_features to native cc_test
This commit pops `transitive_features` off of kwargs when building a
native cc_test. When the `cc_test = cc_test_and_cc_enclave_test` alias
is used within BUILD files, `transitive_features` is a valid attribute
on the cc_enclave_test rule but not the cc_test rule. This change
allows `transitive_features` to be passed to the compound rule without
breaking the cc_test build.

PiperOrigin-RevId: 310959695
Change-Id: Icf258cfa983cf533b9f834eaf8bc5d3abd57866c