Add an alignment step to embed_enclaves
The upstream linux kernel driver for SGX requires that enclaves
are page-aligned, so we align the section in the elf file so that
the elf load can be page-aligned and automatically make the enclave
sections page-aligned.

This change uses 2 objcopy invocations because GNU-objcopy does not
support setting the alignment of an added section. This change
requires the binutils in the toolchain be upgraded to at least 2.33
in order to include the --set-section-alignment flag.

In application_wrapper_enclave_test:
test_enclave offset before: 0309582c (not page aligned)
test_enclave offset after:  03096000 (page aligned)

PiperOrigin-RevId: 324113980
Change-Id: I9c594d321033c808357ccddf5212299a24155713