Change Status proto methods to free functions
The old methods are now deprecated. In order to avoid a dependency cycle
between //asylo/util:status and //asylo/util:status_helpers, the
implementation of proto conversions has to remain in status.cc. The new
functions simply call the deprecated methods.

These new functions are currently untested because they just wrap calls
to the deprecated methods. When those methods are removed and their
implementations are moved into status_helpers, the tests will be moved
as well.

This change is part of a project to remove asylo::Status in favor of
absl::Status.
PiperOrigin-RevId: 356381617
Change-Id: I33809171c4d74d0b51b648e01edcbfda2d006159

PiperOrigin-RevId: 356853338
Change-Id: I3b12a50bf5a080953f7f5835e4099a69014d922d