Add ConvertStatus() template function
The new ConvertStatus() template function should be able to convert a
status from any one status-like to type to any other. It should be used
in place of other existing conversion methods on asylo::Status, except
for converting between asylo::Status and absl::Status.

This change also deprecates the generic status conversion methods on
asylo::Status.

This change is part of a project to remove asylo::Status in favor of
absl::Status.
PiperOrigin-RevId: 356636905
Change-Id: Iec65c696a8b67871eea6ee08cae113694d5982ab