Add new value access interface to StatusOr
Values held in StatusOr objects can now be accessed through dereference
operators if the StatusOr is OK. The behavior of the dereference
operators is undefined if the StatusOr is not OK.

A new value() method can also be used to access the contained value.
Unlike the dereference operators, value()'s behavior is defined if the
StatusOr is not OK.

The new methods mirror those of absl::StatusOr and std::optional.

This change is part of a project to remove asylo::Status in favor of
absl::Status.
PiperOrigin-RevId: 357283749
Change-Id: I54ae139ebf609425a2c0215a28948d08cb294782