Make type conversions for syscalls fail closed
Previously, many of the type conversion functions would fail open,
returning a good value when the input is unexpected. Adjust all
functions so that they instead return absl::nullopt if the input
cannot be converted to a known good output.

This change necessitates a scrub of all places where Asylo converts
between the linux and internal type spaces to ensure conversion errors
are handled appropriately.

PiperOrigin-RevId: 364685627
Change-Id: I483b4eb0d6421276e264d0feb164f08f5d0f4762