Fail type conversions on unexpected input bits
If a bitfield is passed to a type conversion function, but we encounter
an unexpected bit, fail the conversion by default. This leaves it to the
syscall layer to deal with explicitly masking off unexpected bits if
they are ok to ignore.

Add a flag allowing unexpected bits to be ignored for cases where the
code knows it's ok to throw out unhandled bits.

PiperOrigin-RevId: 365092363
Change-Id: I44e54f46ca3fcfb1a1b4bd37ef0160ea6b702aba