Add a formal EKEP analysis and discussion.
This directory contains a formal analysis of the EKEP protocol in
ProVerif, showing that this protocol satisfies useful security
properties.

This directory also contains some useful tools for writing and testing
modular ProVerif code, including a script for using the C Preprocessor
to write macros, and a simple unit-test framework.